+++ HlslDecompiler.Tests/ShaderSources/cs_5_0_instruction/structured_minmax.fx
@@ -6,7 +6,7 @@ void main(uint3 sv_dispatchthreadid : SV_DispatchThreadID)
 {
 	int r0;
 	r0 = keys[sv_dispatchthreadid.x];
-	InterlockedMin(bounds[0], r0.x);
-	InterlockedMax(bounds[1], r0.x);
+	InterlockedMin(bounds[0], (uint)r0.x);
+	InterlockedMax(bounds[1], (uint)r0.x);
 	InterlockedXor(bounds[2], r0.x);
 }
